first complete rule deletion module
